+++ scripts/question_answering/utils.py
@@ -16,12 +16,12 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-import argparse
 import os
 
 import inspect
 import logging
-from mxnet import nd
+import mxnet as mx
+from mxnet import nd, gluon
 
 
 def logging_config(folder=None, name=None, level=logging.DEBUG, console_level=logging.INFO,
@@ -72,7 +72,6 @@ def logging_config(folder=None, name=None, level=logging.DEBUG, console_level=lo
         logging.root.addHandler(logconsole)
 
     return folder
-
 
 
 def get_very_negative_number():
@@ -325,3 +324,54 @@ def replace_masked_values(F, tensor, mask, replace_with):
     one_minus_mask = 1.0 - mask
     values_to_add = replace_with * one_minus_mask
     return F.broadcast_add(F.broadcast_mul(tensor, mask), values_to_add)
+
+
+class PolyakAveraging:
+    """Class to do Polyak averaging based on this paper
+    http://www.meyn.ece.ufl.edu/archive/spm_files/Courses/ECE555-2011/555media/poljud92.pdf"""
+    def __init__(self, params, decay):
+        self._params = params
+        self._decay = decay
+
+        self._polyak_params_dict = gluon.ParameterDict()
+
+        for param in self._params.values():
+            polyak_param = self._polyak_params_dict.get(param.name, shape=param.shape)
+            polyak_param.initialize(mx.init.Constant(self._param_data_to_cpu(param)), ctx=mx.cpu())
+
+    def update(self):
+        """
+        Updates currently held saved parameters with current state of network.
+
+        All calculations for this average occur on the cpu context.
+        """
+        for param in self._params.values():
+            polyak_param = self._polyak_params_dict.get(param.name)
+            polyak_param.set_data(
+                (1 - self._decay) * self._param_data_to_cpu(param) +
+                self._decay * polyak_param.data(mx.cpu()))
+
+    def get_params(self):
+        """ Provides averaged parameters
+
+        Returns
+        -------
+        gluon.ParameterDict
+            Averaged parameters
+        """
+        return self._polyak_params_dict
+
+    def _param_data_to_cpu(self, param):
+        """Returns a copy (on CPU context) of the data held in some context of given parameter.
+
+        Parameters
+        ----------
+        param: gluon.Parameter
+            Parameter's whose data needs to be copied.
+
+        Returns
+        -------
+        NDArray
+            Copy of data on CPU context.
+        """
+        return param.list_data()[0].copyto(mx.cpu())
